requirements.txt
PyQt5
python-binance
openai
pyqtgraph
pandas
